update transaction table

+++ src/main/java/com/manisha/khatabook/dao/models/response/GetBalanceResponse.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Builder
 @Data
 public class GetBalanceResponse {
-    private int debit;
-    private int credit;
+    private double debit;
+    private double credit;
 }
